febio2: Modified FEElasticMixture and FEUncoupledElasticMixture to properly combine material axes information from parent material and components. Simplified FEContinuousFiberDistribution and FEContinuousFiberDistributionUC to account for this new capability in solid mixtures. Updated serialization in FEElasticMixtureMaterialPoint. Created default settings for fiber orientation in FEFiberExpPow and FEFiberExpPowUncoupled.
git-svn-id: https://gforge.sci.utah.edu/svn/MRLProjects/FEBio2@6235 b505bdfe-9d5c-0410-8600-697d5fc8e487

+++ FEBioMech/FEUncoupledElasticMixture.h
@@ -20,7 +20,7 @@ public:
 	FEMaterialPoint* CreateMaterialPointData();
 
 	// return number of materials
-	int Materials() const { return m_pMat.size(); }
+	int Materials() const { return (int)m_pMat.size(); }
 
 	// return a material component
 	FEUncoupledMaterial* GetMaterial(int i) { return m_pMat[i]; }
@@ -36,7 +36,7 @@ public:
 
 public:
 	//! get number of material properties
-	int Properties() { return m_pMat.size(); }
+	int Properties() { return (int)m_pMat.size(); }
 
 	//! return a material property
 	FECoreBase* GetProperty(int n) { return m_pMat[n]; }
